Validate that column names and struct attributes match.
This is a strictly stronger test than requiring that struct attributes
and column names are the same length, so we can remove that test.

We don't need to check that the struct attribute names are unique (which
is oddly enough not enforced by Ruby) because testing that the
attributes and columns match along with testing the uniqueness of the
column names transitively guarantees that the struct attribute names are
unique.