Add ActiveRecord::Base#run_sql.
This new API method allow execution of arbitrary SQL without implying
that you care about the result set. While at some level that makes it
feel like it's just a substitute for `#value_from_sql` and friends with
different code reading implications, it also as a different functional
value: for modification queries it returns the number of tuples
modified.

In addition it guarantees that the result set isn't turned into an
`ActiveRecord::Result` (to avoid overhead) and that any manual memory
allocation is handled (to avoid problems with `Connection#execute` on
PostgreSQL, for example.

Like all of the other extensions to `ActiveRecord::Base` this method
guarantees that no query caching occurs.